feat(storage): make reactive data storage implementation permanent
- Remove ENABLE_REACTIVE_DATA_STORAGE feature flag from utils/featureFlags.ts
- Update Providers.tsx to always use ReactivePanelStoreProvider
- Replace page components with reactive implementations:
  - apps/app/src/app/(home)/page.tsx
  - apps/app/src/app/panel/[panel]/page.tsx
  - apps/app/src/app/panel/[panel]/view/[view]/page.tsx
- Update storage factory to always use reactive storage for API mode
- Simplify WorklistNavigation to use only reactive implementation
- Delete obsolete files:
  - ReactiveComponentSelector.tsx
  - page-reactive.tsx files (home, panel, view)
  - use-panel-store.tsx and its integration test
- Remove all feature flag conditional logic throughout codebase

BREAKING CHANGE: ENABLE_REACTIVE_DATA_STORAGE feature flag removed, reactive storage is now the default implementation
