+++ apps/app/src/app/panel/[panel]/page.tsx
@@ -1,22 +1,26 @@
 "use client";
 
-import { isFeatureEnabled } from '@/utils/featureFlags'
-import ReactiveWorklistPage from './page-reactive'
 import WorklistFooter from "@/app/panel/[panel]/components/WorklistFooter";
 import WorklistNavigation from "@/app/panel/[panel]/components/WorklistNavigation";
-import WorklistTable from "@/app/panel/[panel]/components/WorklistTable";
+import { VirtualizedWorklistTable } from "@/app/panel/[panel]/components/WorklistVirtualizedTable";
 import WorklistToolbar from "@/app/panel/[panel]/components/WorklistToolbar";
+import { useAuthentication } from "@/hooks/use-authentication";
 import { useColumnCreator } from "@/hooks/use-column-creator";
-import { type WorklistPatient, type WorklistTask, useMedplumStore } from "@/hooks/use-medplum-store";
-import { usePanelStore } from "@/hooks/use-panel-store";
+import { useMedplumStore } from "@/hooks/use-medplum-store";
+import { useReactivePanel } from "@/hooks/use-reactive-data";
+import { useReactivePanelStore } from "@/hooks/use-reactive-panel-store";
 import { useSearch } from "@/hooks/use-search";
 import { arrayMove } from "@/lib/utils";
-import type { ColumnDefinition, Filter, PanelDefinition, SortConfig, ViewDefinition, WorklistDefinition } from "@/types/worklist";
+import type { ColumnDefinition, Filter, SortConfig, ViewDefinition, WorklistDefinition } from "@/types/worklist";
 import type { DragEndEvent } from "@dnd-kit/core";
-import { useParams, useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { AddIngestionModal } from "./components/AddIngestionModal";
 import { Loader2 } from "lucide-react";
+import { useParams, useRouter } from "next/navigation";
+import { useEffect, useState, useCallback } from "react";
+import { AddIngestionModal } from "./components/AddIngestionModal";
+import { useDrawer } from "@/contexts/DrawerContext";
+import { TaskDetails } from "./components/TaskDetails";
+import { PatientContext } from "./components/PatientContext";
+import type { WorklistPatient, WorklistTask } from "@/hooks/use-medplum-store";
 
 interface TableFilter {
   key: string;
@@ -24,27 +28,26 @@ interface TableFilter {
 }
 
 export default function WorklistPage() {
-  const isReactiveEnabled = isFeatureEnabled('ENABLE_REACTIVE_DATA_STORAGE')
-
-  // If reactive is enabled, use the reactive component
-  if (isReactiveEnabled) {
-    return <ReactiveWorklistPage />
-  }
-
-  // Original implementation for when reactive is disabled
   const params = useParams();
   const panelId = params.panel as string;
   const [currentView, setCurrentView] = useState<'patient' | 'task'>('patient');
   const [isAddingIngestionSource, setIsAddingIngestionSource] = useState(false);
   const [tableFilters, setTableFilters] = useState<TableFilter[]>([]);
   const [sortConfig, setSortConfig] = useState<SortConfig | undefined>(undefined);
+  const [selectedRows, setSelectedRows] = useState<string[]>([]);
 
+  const { name: currentUserName } = useAuthentication();
   const { patients, tasks, toggleTaskOwner, isLoading: isMedplumLoading } = useMedplumStore();
-  const { getPanel, updatePanel, addView } = usePanelStore();
+  const { updatePanel, addView, updateColumn } = useReactivePanelStore();
+  const { panel, isLoading: isPanelLoading, error: panelError } = useReactivePanel(panelId);
+  const { openDrawer } = useDrawer();
+
   const router = useRouter();
 
-  // Get panel data
-  const panel = getPanel(panelId);
+  // Get columns for current view type
+  const columns = currentView === 'patient'
+    ? (panel?.patientViewColumns || [])
+    : (panel?.taskViewColumns || []);
 
   // Set table data based on current view
   const tableData = currentView === 'patient' ? patients : tasks;
@@ -63,10 +66,10 @@ export default function WorklistPage() {
 
   // Handle panel not found
   useEffect(() => {
-    if (!panel) {
+    if (!isPanelLoading && !panel && !panelError) {
       router.push('/');
     }
-  }, [panel, router]);
+  }, [isPanelLoading, panel, panelError, router]);
 
   const onColumnChange = async (column: WorklistDefinition | ViewDefinition) => {
     if (!panel) {
@@ -99,7 +102,7 @@ export default function WorklistPage() {
     }
 
     try {
-      const newView = await addView(panel.id, {
+      const newView = await addView?.(panel.id, {
         title: "New View",
         filters: panel.filters,
         columns: currentView === 'patient' ? panel.patientViewColumns : panel.taskViewColumns,
@@ -121,41 +124,19 @@ export default function WorklistPage() {
     }
 
     try {
-      await updatePanel(panel.id, { title: newTitle });
+      await updatePanel?.(panel.id, { title: newTitle });
     } catch (error) {
       console.error('Failed to update panel title:', error);
     }
   }
 
   const onColumnUpdate = async (updates: Partial<ColumnDefinition>) => {
-    if (!panel) {
+    if (!panel || !updates.id) {
       return;
     }
 
-    const newPanel = {
-      ...panel,
-      taskViewColumns: panel.taskViewColumns.map(column => {
-        if (column.id === updates.id) {
-          return {
-            ...column,
-            ...updates,
-          }
-        }
-        return column;
-      }),
-      patientViewColumns: panel.patientViewColumns.map(column => {
-        if (column.id === updates.id) {
-          return {
-            ...column,
-            ...updates,
-          }
-        }
-        return column;
-      }),
-    }
-
     try {
-      await updatePanel(panel.id, newPanel);
+      await updateColumn?.(panel.id, updates.id, updates);
     } catch (error) {
       console.error('Failed to update column:', error);
     }
@@ -165,6 +146,9 @@ export default function WorklistPage() {
     if (!panel) {
       return;
     }
+
+    // ✅ OPTIMISTIC UPDATE: Apply filters immediately to UI
+    setTableFilters(newTableFilters);
 
     // Convert table filters to view filters
     const newFilters: Filter[] = newTableFilters.map(filter => ({
@@ -176,12 +160,39 @@ export default function WorklistPage() {
     }
 
     try {
-      await updatePanel(panel.id, newPanel);
-      setTableFilters(newTableFilters);
+      await updatePanel?.(panelId, newPanel);
+      // Filters already applied to UI above
     } catch (error) {
       console.error('Failed to update filters:', error);
     }
   }
+
+  const toggleSelectRow = (rowId: string) => {
+    setSelectedRows(prev => prev.includes(rowId) ? prev.filter(id => id !== rowId) : [...prev, rowId]);
+  };
+
+  const toggleSelectAll = () => {
+    setSelectedRows(prev => prev.length === filteredData.length ? [] : filteredData.map(item => item.id));
+  };
+
+  // Centralized row click handler - optimized with useCallback
+  const handleRowClick = useCallback((row: Record<string, any>) => {
+    if (currentView === "task") {
+      openDrawer(
+        <TaskDetails
+          taskData={row as WorklistTask}
+        />,
+        row.description || "Task Details"
+      );
+    } else if (currentView === "patient") {
+      openDrawer(
+        <PatientContext
+          patient={row as WorklistPatient}
+        />,
+        `${row.name} - Patient Details`
+      );
+    }
+  }, [currentView, openDrawer]);
 
   const handleDragEnd = async (event: DragEndEvent) => {
     const { active, over } = event;
@@ -190,7 +201,6 @@ export default function WorklistPage() {
     }
 
     // Find the active column's index and the over column's index
-    const columns = currentView === 'patient' ? panel.patientViewColumns : panel.taskViewColumns;
     const oldIndex = columns.findIndex(col => col.id === active.id);
     const newIndex = columns.findIndex(col => col.id === over.id);
 
@@ -227,91 +237,89 @@ export default function WorklistPage() {
     };
 
     try {
-      await updatePanel(panel.id, newPanel);
+      await updatePanel?.(panel.id, newPanel);
     } catch (error) {
       console.error('Failed to reorder columns:', error);
     }
   }
 
-  if (!panel) {
-    return null;
-  }
+  const isLoading = isPanelLoading || !panel;
 
   return (
     <>
-      {/* Navigation Area */}
-      <div className="navigation-area">
-        <WorklistNavigation
-          panelDefinition={panel!}
-          onNewView={onNewView}
-          onPanelTitleChange={onPanelTitleChange}
-        />
-      </div>
-
-      {/* Toolbar Area */}
-      <div className="toolbar-area">
-        <WorklistToolbar
-          searchTerm={searchTerm}
-          onSearch={setSearchTerm}
-          searchMode={searchMode}
-          onSearchModeChange={setSearchMode}
-          currentView={currentView}
-          setCurrentView={setCurrentView}
-          worklistColumns={currentView === 'patient' ? panel.patientViewColumns : panel.taskViewColumns}
-          onAddColumn={onAddColumn}
-          onColumnVisibilityChange={(columnId, visible) => onColumnUpdate({
-            id: columnId,
-            properties: {
-              display: { visible }
-            }
-          })}
-        />
-      </div>
-
-      {/* Content Area */}
-      <div className="content-area">
-        <div className="table-scroll-container">
-          <WorklistTable
-            isLoading={isMedplumLoading}
-            selectedRows={[]}
-            toggleSelectAll={() => { }}
-            onSortConfigUpdate={setSortConfig}
-            worklistColumns={currentView === 'patient' ? panel.patientViewColumns : panel.taskViewColumns}
-            tableData={filteredData}
-            handlePDFClick={() => { }}
-            handleTaskClick={() => { }}
-            handleRowHover={() => { }}
-            toggleSelectRow={() => { }}
-            handleAssigneeClick={(taskId: string) => toggleTaskOwner(taskId)}
-            setIsAddingIngestionSource={() => setIsAddingIngestionSource(true)}
-            currentView={currentView}
-            handleDragEnd={handleDragEnd}
-            onColumnUpdate={onColumnUpdate}
-            filters={tableFilters}
-            onFiltersChange={onFiltersChange}
-            initialSortConfig={sortConfig ?? null}
-          />
+      {isLoading ? (
+        <div className="flex items-center justify-center h-screen">
+          <Loader2 className="h-8 w-8 text-blue-500 animate-spin mb-2" aria-label="Loading Panel" />
         </div>
-
-        {isAddingIngestionSource && (
-          <AddIngestionModal
-            isOpen={isAddingIngestionSource}
-            onClose={() => setIsAddingIngestionSource(false)}
-            onSelectSource={() => { }}
-            ingestionBots={[]}
-          />
-        )}
-      </div>
-
-      {/* Footer Area */}
-      <div className="footer-area">
-        <WorklistFooter
-          columnsCounter={currentView === 'patient' ? panel.patientViewColumns.length : panel.taskViewColumns.length}
-          rowsCounter={tableData.length}
-          navigateToHome={() => router.push('/')}
-          isAISidebarOpen={false}
-        />
-      </div>
+      ) : (
+        <>
+          <div className="navigation-area">
+            {panel && (
+              <WorklistNavigation panelDefinition={panel} onNewView={onNewView} onPanelTitleChange={onPanelTitleChange} />
+            )}
+          </div>
+          <div className="toolbar-area">
+            <WorklistToolbar
+              key={`${panelId}-${currentView}-${columns.length}`}
+              searchTerm={searchTerm}
+              onSearch={setSearchTerm}
+              searchMode={searchMode}
+              onSearchModeChange={setSearchMode}
+              currentView={currentView}
+              setCurrentView={setCurrentView}
+              worklistColumns={columns}
+              onAddColumn={onAddColumn}
+              onColumnVisibilityChange={(columnId, visible) => onColumnUpdate({
+                id: columnId,
+                properties: {
+                  display: { visible }
+                }
+              })}
+            />
+          </div>
+          <div className="content-area">
+            <div className="table-scroll-container">
+              <VirtualizedWorklistTable
+                isLoading={isMedplumLoading}
+                selectedRows={selectedRows}
+                toggleSelectAll={toggleSelectAll}
+                onSortConfigUpdate={setSortConfig}
+                worklistColumns={columns}
+                tableData={filteredData}
+                handlePDFClick={() => { }}
+                handleTaskClick={() => { }}
+                handleRowHover={() => { }}
+                toggleSelectRow={toggleSelectRow}
+                handleAssigneeClick={(taskId: string) => toggleTaskOwner(taskId)}
+                currentUserName={currentUserName}
+                currentView={currentView}
+                onColumnUpdate={onColumnUpdate}
+                filters={tableFilters}
+                onFiltersChange={onFiltersChange}
+                initialSortConfig={sortConfig ?? null}
+                onRowClick={handleRowClick}
+                handleDragEnd={handleDragEnd}
+              />
+              {isAddingIngestionSource && (
+                <AddIngestionModal
+                  isOpen={isAddingIngestionSource}
+                  onClose={() => setIsAddingIngestionSource(false)}
+                  onSelectSource={() => { }}
+                  ingestionBots={[]}
+                />
+              )}
+            </div>
+          </div>
+          <div className="footer-area">
+            <WorklistFooter
+              columnsCounter={columns.length}
+              rowsCounter={tableData.length}
+              navigateToHome={() => router.push('/')}
+              isAISidebarOpen={false}
+            />
+          </div>
+        </>
+      )}
     </>
   );
 }
