+++ apps/app/src/app/panel/[panel]/view/[view]/page.tsx
@@ -1,20 +1,22 @@
 "use client";
-import { isFeatureEnabled } from '@/utils/featureFlags'
-import ReactiveWorklistViewPage from './page-reactive'
 import WorklistFooter from "@/app/panel/[panel]/components/WorklistFooter";
 import WorklistNavigation from "@/app/panel/[panel]/components/WorklistNavigation";
-import WorklistTable from "@/app/panel/[panel]/components/WorklistTable";
+import { VirtualizedWorklistTable } from "@/app/panel/[panel]/components/WorklistVirtualizedTable";
 import WorklistToolbar from "@/app/panel/[panel]/components/WorklistToolbar";
 import { useColumnCreator } from "@/hooks/use-column-creator";
 import { useMedplumStore } from "@/hooks/use-medplum-store";
-import { usePanelStore } from "@/hooks/use-panel-store";
+import { useReactivePanelStore } from "@/hooks/use-reactive-panel-store";
+import { useReactivePanel, useReactiveView } from "@/hooks/use-reactive-data";
 import { useSearch } from "@/hooks/use-search";
 import { arrayMove } from "@/lib/utils";
 import type { ColumnDefinition, Filter, PanelDefinition, SortConfig, ViewDefinition, WorklistDefinition } from "@/types/worklist";
 import type { DragEndEvent } from "@dnd-kit/core";
-import { Loader2 } from "lucide-react";
 import { useParams, useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
+import { useEffect, useState, useCallback } from "react";
+import { useDrawer } from "@/contexts/DrawerContext";
+import { TaskDetails } from "@/app/panel/[panel]/components/TaskDetails";
+import { PatientContext } from "@/app/panel/[panel]/components/PatientContext";
+import type { WorklistPatient, WorklistTask } from "@/hooks/use-medplum-store";
 
 interface TableFilter {
   key: string;
@@ -22,40 +24,17 @@ interface TableFilter {
 }
 
 export default function WorklistViewPage() {
-  const isReactiveEnabled = isFeatureEnabled('ENABLE_REACTIVE_DATA_STORAGE')
-
-  // If reactive is enabled, use the reactive component
-  if (isReactiveEnabled) {
-    return <ReactiveWorklistViewPage />
-  }
-
-  // Original implementation for when reactive is disabled
   const { patients, tasks, toggleTaskOwner, isLoading: isMedplumLoading } = useMedplumStore();
-  const { getPanel, getView, updatePanel, addView, isLoading: isPanelLoading } = usePanelStore();
+  const { updatePanel, updateView, addView, updateColumn } = useReactivePanelStore();
+  const { openDrawer } = useDrawer();
   const params = useParams();
   const panelId = params.panel as string;
   const viewId = params.view as string;
   const router = useRouter();
 
+  const { panel, isLoading: isPanelLoading, error: panelError } = useReactivePanel(panelId);
+  const { view, isLoading: isViewLoading, error: viewError } = useReactiveView(panelId, viewId);
   const [tableFilters, setTableFilters] = useState<TableFilter[]>([]);
-  const [panel, setPanel] = useState<PanelDefinition | null>(null);
-  const [view, setView] = useState<ViewDefinition | null>(null);
-
-  // Get panel and view data
-  useEffect(() => {
-    const currentPanel = getPanel(panelId);
-    if (currentPanel) {
-      setPanel(currentPanel);
-      const currentView = getView(panelId, viewId);
-      if (currentView) {
-        setView(currentView);
-        setTableFilters(currentView.filters.map((filter: Filter) => ({
-          key: filter.fhirPathFilter[0],
-          value: filter.fhirPathFilter[1],
-        })));
-      }
-    }
-  }, [getPanel, getView, panelId, viewId]);
 
   const searchData = view?.viewType === 'patient' ? patients : tasks;
   // @ts-ignore - Type mismatch between patient/task arrays but useSearch handles both
@@ -66,57 +45,48 @@ export default function WorklistViewPage() {
     : (view?.viewType === 'patient' ? panel?.patientViewColumns ?? [] : panel?.taskViewColumns ?? []);
   const tableData = filteredData ?? [];
 
+  // Set filters from view
+  useEffect(() => {
+    if (view) {
+      setTableFilters(view.filters.map(filter => ({
+        key: filter.fhirPathFilter[0],
+        value: filter.fhirPathFilter[1],
+      })));
+    }
+  }, [view]);
+
   // Handle panel or view not found
   useEffect(() => {
-    if (!isPanelLoading && !panel) {
+    if (!isPanelLoading && !panel && !panelError) {
       router.push('/');
       return;
     }
-    if (!isPanelLoading && !view) {
+    if (!isViewLoading && !view && !viewError) {
       router.push(`/panel/${panelId}`);
       return;
     }
-  }, [isPanelLoading, panel, view, router, panelId]);
+  }, [isPanelLoading, isViewLoading, panel, view, panelError, viewError, router, panelId]);
 
   const onColumnUpdate = async (updates: Partial<ColumnDefinition>) => {
-    if (!view || !panel) {
+    if (!view || !panel || !updates.id) {
       return;
     }
 
-    const panelColumns = view.viewType === 'patient' ? panel.patientViewColumns ?? [] : panel.taskViewColumns ?? []
-
-    const newColumns = updates.properties?.display?.visible
-      ? view.columns.some((col: ColumnDefinition) => col.id === updates.id)
-        ? view.columns
-        : [...view.columns, panelColumns.find((col: ColumnDefinition) => col.id === updates.id)].filter((col): col is ColumnDefinition => col !== undefined)
-      : view.columns.filter((column: ColumnDefinition) => column.id !== updates.id).filter((col): col is ColumnDefinition => col !== undefined)
-
-    const newView = {
-      ...view,
-      columns: newColumns
-    }
-
-    const newPanel = {
-      ...panel,
-      views: panel.views?.map((v: ViewDefinition) => v.id === viewId ? newView : v) ?? [newView]
-    }
-
     try {
-      await updatePanel(panel.id, newPanel);
-      setView(newView);
-      setPanel(newPanel);
+      // Update the column in the panel first
+      await updateColumn?.(panelId, updates.id, updates);
     } catch (error) {
       console.error('Failed to update column:', error);
     }
   }
 
   const onFiltersChange = async (newTableFilters: TableFilter[]) => {
-    if (!view || !panel) {
+    if (!view) {
       return;
     }
 
     // Convert table filters to view filters
-    const newFilters: Filter[] = newTableFilters.map((filter: TableFilter) => ({
+    const newFilters: Filter[] = newTableFilters.map(filter => ({
       fhirPathFilter: [filter.key, filter.value]
     }));
     const newView = {
@@ -124,16 +94,9 @@ export default function WorklistViewPage() {
       filters: newFilters,
     }
 
-    const newPanel = {
-      ...panel,
-      views: panel.views?.map((v: ViewDefinition) => v.id === viewId ? newView : v) ?? [newView]
-    }
-
     try {
-      await updatePanel(panel.id, newPanel);
+      await updateView?.(panelId, viewId, newView);
       setTableFilters(newTableFilters);
-      setView(newView);
-      setPanel(newPanel);
     } catch (error) {
       console.error('Failed to update filters:', error);
     }
@@ -141,13 +104,13 @@ export default function WorklistViewPage() {
 
   const handleDragEnd = async (event: DragEndEvent) => {
     const { active, over } = event;
-    if (!over || active.id === over.id || !view || !panel) {
+    if (!over || active.id === over.id || !view) {
       return;
     }
 
     // Find the active column's index and the over column's index
-    const oldIndex = columns.findIndex((col: ColumnDefinition) => col.id === active.id);
-    const newIndex = columns.findIndex((col: ColumnDefinition) => col.id === over.id);
+    const oldIndex = columns.findIndex(col => col.id === active.id);
+    const newIndex = columns.findIndex(col => col.id === over.id);
 
     if (oldIndex === -1 || newIndex === -1) {
       return;
@@ -162,22 +125,15 @@ export default function WorklistViewPage() {
       columns: reorderedColumns,
     };
 
-    const newPanel = {
-      ...panel,
-      views: panel.views?.map((v: ViewDefinition) => v.id === viewId ? newView : v) ?? [newView]
-    }
-
     try {
-      await updatePanel(panel.id, newPanel);
-      setView(newView);
-      setPanel(newPanel);
+      await updateView?.(panelId, viewId, newView);
     } catch (error) {
       console.error('Failed to reorder columns:', error);
     }
   }
 
   const onColumnChange = async (column: ViewDefinition | WorklistDefinition) => {
-    if (!view || !panel) {
+    if (!view) {
       return;
     }
 
@@ -186,15 +142,8 @@ export default function WorklistViewPage() {
       ...column,
     }
 
-    const newPanel = {
-      ...panel,
-      views: panel.views?.map((v: ViewDefinition) => v.id === viewId ? newView : v) ?? [newView]
-    }
-
     try {
-      await updatePanel(panel.id, newPanel);
-      setView(newView);
-      setPanel(newPanel);
+      await updateView?.(panelId, viewId, newView);
     } catch (error) {
       console.error('Failed to update view:', error);
     }
@@ -214,7 +163,7 @@ export default function WorklistViewPage() {
     }
 
     try {
-      const newView = await addView(panel.id, {
+      const newView = await addView?.(panelId, {
         title: "New View",
         filters: view?.filters ?? panel.filters,
         columns: view?.viewType === 'patient' ? panel.patientViewColumns : panel.taskViewColumns,
@@ -231,106 +180,109 @@ export default function WorklistViewPage() {
   }
 
   const onSortConfigUpdate = async (sortConfig: SortConfig | undefined) => {
-    if (!view || !panel) {
+    if (!view) {
       return;
     }
     const newView = {
       ...view,
       sortConfig: sortConfig ? [sortConfig] : [],
     }
-    const newPanel = {
-      ...panel,
-      views: panel.views?.map((v: ViewDefinition) => v.id === viewId ? newView : v) ?? [newView]
-    }
-    await updatePanel(panel.id, newPanel);
-    setView(newView);
-    setPanel(newPanel);
+    await updateView?.(panelId, viewId, newView);
   }
 
+  // Centralized row click handler - optimized with useCallback
+  const handleRowClick = useCallback((row: Record<string, any>) => {
+    if (view?.viewType === "task") {
+      openDrawer(
+        <TaskDetails
+          taskData={row as WorklistTask}
+        />,
+        row.description || "Task Details"
+      );
+    } else if (view?.viewType === "patient") {
+      openDrawer(
+        <PatientContext
+          patient={row as WorklistPatient}
+        />,
+        `${row.name} - Patient Details`
+      );
+    }
+  }, [view?.viewType, openDrawer]);
+
   const onViewTitleChange = async (newTitle: string) => {
-    if (!view || !panel) {
-      console.log("viewDefinition not found");
+    if (!view) {
       return;
     }
 
     try {
-      const newView = {
-        ...view,
-        title: newTitle,
-      }
-      const newPanel = {
-        ...panel,
-        views: panel.views?.map((v: ViewDefinition) => v.id === viewId ? newView : v) ?? [newView]
-      }
-      await updatePanel(panel.id, newPanel);
-      setView(newView);
-      setPanel(newPanel);
+      await updateView?.(panelId, viewId, { title: newTitle });
     } catch (error) {
       console.error('Failed to update view title:', error);
     }
   };
 
-  if (!panel || !view) {
-    return null;
-  }
+  const isLoading = isPanelLoading || isViewLoading || !panel || !view;
 
   return (
     <>
-      {/* Navigation Area */}
-      <div className="navigation-area">
-        <WorklistNavigation panelDefinition={panel} selectedViewId={viewId} onNewView={onNewView} onViewTitleChange={onViewTitleChange} />
-      </div>
-
-      {/* Toolbar Area */}
-      <div className="toolbar-area">
-        <WorklistToolbar
-          searchTerm={searchTerm}
-          onSearch={setSearchTerm}
-          searchMode={searchMode}
-          onSearchModeChange={setSearchMode}
-          currentView={view?.viewType}
-          setCurrentView={() => { }}
-          worklistColumns={columns}
-          onAddColumn={onAddColumn}
-          onColumnVisibilityChange={(columnId, visible) => onColumnUpdate({ id: columnId, properties: { display: { visible } } })}
-        />
-      </div>
-
-      {/* Content Area */}
-      <div className="content-area">
-        <div className="table-scroll-container">
-          <WorklistTable
-            isLoading={isMedplumLoading}
-            selectedRows={[]}
-            toggleSelectAll={() => { }}
-            worklistColumns={columns}
-            onSortConfigUpdate={onSortConfigUpdate}
-            tableData={filteredData}
-            handlePDFClick={() => { }}
-            handleTaskClick={() => { }}
-            handleRowHover={() => { }}
-            toggleSelectRow={() => { }}
-            handleAssigneeClick={(taskId: string) => toggleTaskOwner(taskId)}
-            setIsAddingIngestionSource={() => { }}
-            currentView={view?.viewType ?? 'patient'}
-            handleDragEnd={handleDragEnd}
-            onColumnUpdate={onColumnUpdate}
-            filters={tableFilters}
-            onFiltersChange={onFiltersChange}
-            initialSortConfig={view?.sortConfig?.[0] ?? null}
-          />
+      {isLoading ? (
+        <div className="flex items-center justify-center h-screen">
+          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-gray-900" />
         </div>
-      </div>
-
-      {/* Footer Area */}
-      <div className="footer-area">
-        <WorklistFooter
-          columnsCounter={columns.length}
-          rowsCounter={tableData.length}
-          navigateToHome={() => router.push('/')}
-          isAISidebarOpen={false}
-        />
-      </div>
+      ) : (
+        <>
+          <div className="navigation-area">
+            {panel && (
+              <WorklistNavigation panelDefinition={panel} selectedViewId={viewId} onNewView={onNewView} onViewTitleChange={onViewTitleChange} />
+            )}
+          </div>
+          <div className="toolbar-area">
+            <WorklistToolbar
+              searchTerm={searchTerm}
+              onSearch={setSearchTerm}
+              searchMode={searchMode}
+              onSearchModeChange={setSearchMode}
+              currentView={view?.viewType}
+              setCurrentView={() => { }}
+              worklistColumns={columns}
+              onAddColumn={onAddColumn}
+              onColumnVisibilityChange={(columnId, visible) => onColumnUpdate({ id: columnId, properties: { display: { visible } } })}
+            />
+          </div>
+          <div className="content-area">
+            <div className="table-scroll-container">
+              <VirtualizedWorklistTable
+                isLoading={isMedplumLoading}
+                selectedRows={[]}
+                toggleSelectAll={() => { }}
+                worklistColumns={columns}
+                onSortConfigUpdate={onSortConfigUpdate}
+                tableData={filteredData}
+                handlePDFClick={() => { }}
+                handleTaskClick={() => { }}
+                handleRowHover={() => { }}
+                toggleSelectRow={() => { }}
+                handleAssigneeClick={(taskId: string) => toggleTaskOwner(taskId)}
+                currentView={view?.viewType ?? 'patient'}
+                onColumnUpdate={onColumnUpdate}
+                filters={tableFilters}
+                onFiltersChange={onFiltersChange}
+                initialSortConfig={view?.sortConfig?.[0] ?? null}
+                onRowClick={handleRowClick}
+                handleDragEnd={handleDragEnd}
+              />
+            </div>
+          </div>
+          <div className="footer-area">
+            <WorklistFooter
+              columnsCounter={columns.length}
+              rowsCounter={tableData.length}
+              navigateToHome={() => router.push('/')}
+              isAISidebarOpen={false}
+            />
+          </div>
+        </>
+      )}
     </>
   );
 }
