+++ apps/app/src/components/Providers.tsx
@@ -1,9 +1,7 @@
 'use client'
 
 import { MedplumProvider } from '@/contexts/medplum-context';
-import { PanelStoreProvider } from '@/hooks/use-panel-store';
 import { ReactivePanelStoreProvider } from '@/hooks/use-reactive-panel-store';
-import { isFeatureEnabled } from '@/utils/featureFlags';
 import { StytchB2BProvider } from '@stytch/nextjs/b2b'
 import { createStytchB2BUIClient } from '@stytch/nextjs/b2b/ui'
 import { CookiesProvider } from 'react-cookie';
@@ -26,14 +24,14 @@ export default function Providers({ children }: { children: React.ReactNode }) {
     };
     fetchConfig();
   }, []);
-  
+
   const stytch = useMemo(() => {
 
     if (!authStytchPublicToken || !authCookiesAllowedDomain) {
       console.log('AUTH_STYTCH_PUBLIC_TOKEN or AUTH_COOKIES_ALLOWED_DOMAIN is not available in runtime config');
       return undefined
     }
-    
+
     return createStytchB2BUIClient(authStytchPublicToken, {
       cookieOptions: {
         opaqueTokenCookieName: `stytch_session`,
@@ -41,7 +39,7 @@ export default function Providers({ children }: { children: React.ReactNode }) {
         availableToSubdomains: true
       }
     });
-  }, [authStytchPublicToken, authCookiesAllowedDomain ]);
+  }, [authStytchPublicToken, authCookiesAllowedDomain]);
 
   // Don't render if stytch is not available
   if (!stytch) {
@@ -54,13 +52,10 @@ export default function Providers({ children }: { children: React.ReactNode }) {
     );
   }
 
-  const isReactiveEnabled = isFeatureEnabled('ENABLE_REACTIVE_DATA_STORAGE');
-
   // TODO we want to start using Stytch with nextjs auth
   return (
     <StytchB2BProvider stytch={stytch}>
       <CookiesProvider>
-
         <AuthenticationGuard loadingComponent={
           <div className="h-screen w-full flex items-center justify-center">
             <Loader2 className="h-8 w-8 text-blue-500 animate-spin mb-2" aria-label="Authenticating..." />
@@ -68,15 +63,9 @@ export default function Providers({ children }: { children: React.ReactNode }) {
         }>
           <AuthenticationStoreProvider>
             <MedplumProvider>
-              {isReactiveEnabled ? (
-                <ReactivePanelStoreProvider>
-                  {children}
-                </ReactivePanelStoreProvider>
-              ) : (
-                <PanelStoreProvider>
-                  {children}
-                </PanelStoreProvider>
-              )}
+              <ReactivePanelStoreProvider>
+                {children}
+              </ReactivePanelStoreProvider>
             </MedplumProvider>
           </AuthenticationStoreProvider>
         </AuthenticationGuard>
